extracting the viewpager scrolling from the landing-strip
