+++ landing-strip/core/src/main/java/com/novoda/landingstrip/LandingStrip.java
@@ -69,73 +69,9 @@ public class LandingStrip extends HorizontalScrollView {
 
     public void setViewPager(ViewPager viewPager, PagerAdapter pagerAdapter, TabSetterUpper tabSetterUpper) {
         this.viewPager = viewPager;
-        viewPager.setOnPageChangeListener(pageListener);
+        viewPager.setOnPageChangeListener(new Foo(state, tabsContainer, indicatorCoordinatesCalculator, this));
         notifyDataSetChanged(pagerAdapter, tabSetterUpper);
     }
-
-    private final ViewPager.SimpleOnPageChangeListener pageListener = new ViewPager.SimpleOnPageChangeListener() {
-
-        @SuppressWarnings("checkstyle:visibilityModifierCheck") // the scope is super small, setter and getter would be overkill
-        private boolean firstTimeAccessed = true;
-
-        @Override
-        public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-            if (firstTimeAccessed) {
-                setSelected(position);
-                firstTimeAccessed = false;
-            }
-            state.updatePosition(position);
-            state.updateOffset(positionOffset);
-
-            int offset = getHorizontalScrollOffset(position, positionOffset);
-            scrollToChild(position, offset);
-            state.getDelegateOnPageListener().onPageScrolled(position, positionOffset, positionOffsetPixels);
-        }
-
-        private int getHorizontalScrollOffset(int position, float swipePositionOffset) {
-            int tabWidth = tabsContainer.getChildAt(position).getWidth();
-            return Math.round(swipePositionOffset * tabWidth);
-        }
-
-        private void scrollToChild(int position, int offset) {
-            Coordinates indicatorCoordinates = indicatorCoordinatesCalculator.calculateIndicatorCoordinates(state, tabsContainer);
-            float newScrollX = calculateScrollOffset(position, offset, indicatorCoordinates);
-
-            state.updateIndicatorCoordinates(indicatorCoordinates);
-
-            scrollTo((int) newScrollX, 0);
-        }
-
-        private float calculateScrollOffset(int position, int offset, Coordinates indicatorCoordinates) {
-            float newScrollX = tabsContainer.getChildAt(position).getLeft() + offset;
-            newScrollX -= getWidth() / 2;
-            newScrollX += ((indicatorCoordinates.getEnd() - indicatorCoordinates.getStart()) / 2f);
-            return newScrollX;
-        }
-
-        @Override
-        public void onPageSelected(int position) {
-            setSelected(position);
-            state.getDelegateOnPageListener().onPageSelected(position);
-        }
-
-        private void setSelected(int position) {
-            int childCount = tabsContainer.getChildCount();
-            for (int index = 0; index < childCount; index++) {
-                if (index == position) {
-                    tabsContainer.getChildAt(position).setSelected(true);
-                } else {
-                    tabsContainer.getChildAt(index).setSelected(false);
-                }
-            }
-        }
-
-        @Override
-        public void onPageScrollStateChanged(int changedState) {
-            state.getDelegateOnPageListener().onPageScrollStateChanged(changedState);
-        }
-
-    };
 
     @Override
     public void scrollTo(int x, int y) {
